+++ fluidmcp/frontend/src/pages/MCPInspector.tsx
@@ -1,270 +1,16 @@
-import React, { useState, useEffect, useRef } from "react";
+import { useState, useEffect, useRef } from "react";
 import { Navbar } from "@/components/Navbar";
 import { apiClient } from "@/services/api";
-import { type SavedRequest, loadSavedRequests, saveRequest, deleteSavedRequest, renameSavedRequest } from "@/lib/saved-requests";
-import { JsonSchemaForm } from '../components/form/JsonSchemaForm';
-import { ToolResult } from '../components/result/ToolResult';
-import { JsonResultView } from '../components/result/JsonResultView';
-import { McpContentView } from '../components/result/McpContentView';
-import { WidgetSandbox } from '../components/WidgetSandbox';
-import { PanelGroup, Panel , PanelResizeHandle} from 'react-resizable-panels';
+import { type SavedRequest, loadSavedRequests, renameSavedRequest } from "@/lib/saved-requests";
+import { ServerListPanel } from '../components/inspector/ServerListPanel';
+import { PanelGroup, Panel, PanelResizeHandle } from 'react-resizable-panels';
 import { LogsPanel } from '../components/inspector/LogsPanel';
+import { AddServerModal } from '../components/inspector/AddServerModal';
 import { ResourcesPanel } from '../components/inspector/ResourcesPanel';
 import { PromptsPanel } from '../components/inspector/PromptsPanel';
-import { AddServerModal } from '../components/inspector/AddServerModal';
-
-// Compact collapsible result bubble for chat mode
-function ChatResultBubble({ result, initialView = "formatted", hideTabSwitcher = false }: { result: unknown; initialView?: "formatted" | "raw"; hideTabSwitcher?: boolean }) {
-  const [expanded, setExpanded] = useState(true);
-  const [viewMode, setViewMode] = useState<"formatted" | "raw">(initialView);
-  useEffect(() => { setViewMode(initialView); }, [initialView]);
-  const isMcp = typeof result === "object" && result !== null &&
-    "content" in result && Array.isArray((result as any).content);
-  const isMcpArray = Array.isArray(result) && result.length > 0 &&
-    (result as any[]).every((i: any) => typeof i === "object" && i !== null && "type" in i);
-  const preview = (() => {
-    if (isMcp) {
-      const texts = ((result as any).content || []).filter((c: any) => c.type === "text" && c.text);
-      if (texts.length > 0) return String(texts[0].text).slice(0, 80);
-      return `[${(result as any).content?.length || 0} items]`;
-    }
-    if (isMcpArray) {
-      const texts = (result as any[]).filter((c: any) => c.type === "text" && c.text);
-      if (texts.length > 0) return String(texts[0].text).slice(0, 80);
-      return `[${(result as any[]).length} items]`;
-    }
-    if (typeof result === "object" && result !== null) return `{${Object.keys(result as object).length} keys}`;
-    return String(result).slice(0, 60);
-  })();
-
-  const tabStyle = (active: boolean) => ({
-    padding: "0.2rem 0.5rem", borderRadius: "0.25rem", border: "none",
-    fontSize: "0.75rem", fontWeight: 500 as const, cursor: "pointer",
-    background: active ? "#fff" : "transparent",
-    color: active ? "#000" : "rgba(255,255,255,0.6)"
-  });
-
-  return (
-    <div style={{ fontSize: "0.8rem" }}>
-      <div style={{ display: "flex", alignItems: "center", gap: "0.5rem" }}>
-        <button
-          onClick={() => setExpanded(v => !v)}
-          style={{
-            background: "none", border: "none", cursor: "pointer",
-            color: "#22c55e", fontWeight: 600, padding: "0.25rem 0",
-            fontSize: "0.8rem", display: "flex", alignItems: "center", gap: "0.3rem"
-          }}
-        >
-          {expanded ? "▼" : "▶"} Result:{!expanded && <span style={{ color: "rgba(255,255,255,0.6)", fontWeight: 400, marginLeft: "0.3rem" }}>{preview}</span>}
-        </button>
-        {expanded && !hideTabSwitcher && (
-          <div style={{ display: "flex", alignItems: "center", gap: "0.4rem" }}>
-            <div style={{ display: "flex", background: "rgba(0,0,0,0.3)", borderRadius: "0.25rem", padding: "0.15rem", gap: "0.15rem" }}>
-              <button style={tabStyle(viewMode === "formatted")} onClick={() => setViewMode("formatted")}>Formatted</button>
-              <button style={tabStyle(viewMode === "raw")} onClick={() => setViewMode("raw")}>Raw JSON</button>
-            </div>
-            <button
-              onClick={() => navigator.clipboard.writeText(JSON.stringify(result, null, 2))}
-              style={{ ...tabStyle(false), border: "1px solid rgba(63,63,70,0.5)", borderRadius: "0.25rem" }}
-              title="Copy to clipboard"
-            >
-              Copy
-            </button>
-          </div>
-        )}
-      </div>
-      {expanded && (
-        <div style={{
-          background: "rgba(0,0,0,0.3)",
-          border: "1px solid rgba(63,63,70,0.5)",
-          borderRadius: "0.5rem",
-          padding: "0.75rem",
-          marginTop: "0.25rem",
-          width: "100%", boxSizing: "border-box",
-          maxHeight: "350px",
-          overflowY: "auto",
-        }}>
-          {viewMode === "raw"
-            ? <pre style={{ margin: 0, fontSize: "0.75rem", color: "#e5e7eb", whiteSpace: "pre-wrap", wordBreak: "break-word", fontFamily: "ui-monospace, monospace", width: "100%", boxSizing: "border-box" as const }}>{JSON.stringify(result, null, 2)}</pre>
-            : <div style={{ minWidth: 0, width: "100%", overflow: "hidden" }}>
-                {(isMcp || isMcpArray)
-                  ? <McpContentView content={isMcpArray ? result as any : (result as any).content} />
-                  : <JsonResultView data={result} />
-                }
-              </div>
-          }
-        </div>
-      )}
-    </div>
-  );
-}
-// ── Animated thinking indicator ──────────────────────────────────────────────
-function ThinkingDots() {
-  return (
-    <span style={{ fontStyle: "italic", color: "rgba(255,255,255,0.4)", fontSize: "0.72rem" }}>
-      Thinking
-      <span style={{ display: "inline-flex", gap: "1px" }}>
-        {[0, 1, 2].map(i => (
-          <span key={i} style={{ animation: `thinking-blink 1.4s ease-in-out ${i * 0.2}s infinite` }}>.</span>
-        ))}
-      </span>
-    </span>
-  );
-}
-
-// ── Group flat chat messages into standalone + run blocks ─────────────────────
-type DisplayGroup =
-  | { kind: "standalone"; msg: ChatMessage }
-  | { kind: "run"; runId: string; steps: ChatMessage[]; run?: ExecutionRun }
-
-function groupMessages(messages: ChatMessage[], execHistory: ExecutionRun[]): DisplayGroup[] {
-  const runMap = new Map(execHistory.map(r => [r.runId, r]));
-  const seen = new Set<string>();
-  return messages.reduce<DisplayGroup[]>((acc, msg) => {
-    if (!msg.runId) {
-      acc.push({ kind: "standalone", msg });
-    } else if (!seen.has(msg.runId)) {
-      seen.add(msg.runId);
-      acc.push({
-        kind: "run",
-        runId: msg.runId,
-        steps: messages.filter(m => m.runId === msg.runId),
-        run: runMap.get(msg.runId),
-      });
-    }
-    return acc;
-  }, []);
-}
-
-// ── Timeline block for one agent turn (thinking → tool_call → tool_result) ───
-function ExecutionRunBlock({ steps, run }: { steps: ChatMessage[]; run?: ExecutionRun }) {
-  const [collapsed, setCollapsed] = useState(false);
-  const thinking  = steps.find(s => s.type === "thinking");
-  const toolCall  = steps.find(s => s.type === "tool_call");
-  const toolResult = steps.find(s => s.type === "tool_result");
-  const errorStep = steps.find(s => s.type === "error");
-  const isActive  = !toolResult && !errorStep && !run?.endTime; // run still in progress
-
-  const totalMs    = run?.endTime ? run.endTime - run.startTime : null;
-  const thinkingMs = (toolCall?.perfMark && thinking?.perfMark)
-    ? Math.round(toolCall.perfMark - thinking.perfMark) : null;
-  const toolMs     = (toolResult?.perfMark && toolCall?.perfMark)
-    ? Math.round(toolResult.perfMark - toolCall.perfMark) : null;
-
-  const dot = (color: string) => (
-    <div style={{ width: "7px", height: "7px", borderRadius: "50%", background: color, marginTop: "4px", flexShrink: 0 }} />
-  );
-
-  return (
-    <div style={{ width: "100%", border: "1px solid rgba(63,63,70,0.5)", borderRadius: "10px" }}>
-      {/* Header */}
-      <div
-        onClick={() => setCollapsed(v => !v)}
-        style={{
-          display: "flex", alignItems: "center", gap: "0.5rem",
-          padding: "0.45rem 0.75rem",
-          background: "rgba(255,255,255,0.04)", cursor: "pointer",
-          borderBottom: collapsed ? "none" : "1px solid rgba(63,63,70,0.25)",
-        }}
-      >
-        <span style={{ fontSize: "0.7rem", opacity: 0.45 }}>{collapsed ? "▶" : "▼"}</span>
-        <span style={{ fontSize: "0.8rem", fontWeight: 600 }}>
-          {toolCall ? `🔧 ${toolCall.toolName}` : errorStep ? "❌ Error" : "🤖 Thinking"}
-        </span>
-        {isActive && <ThinkingDots />}
-        {totalMs !== null && (
-          <span style={{ marginLeft: "auto", fontSize: "0.7rem", color: "rgba(99,102,241,0.7)", fontFamily: "monospace" }}>
-            {totalMs}ms
-          </span>
-        )}
-      </div>
-
-      {!collapsed && (
-        <div style={{ padding: "0.6rem 0.75rem", display: "flex", flexDirection: "column", gap: "0.55rem" }}>
-          {/* Thinking step */}
-          {thinking && (
-            <div style={{ display: "flex", gap: "0.6rem" }}>
-              {dot("#3b82f6")}
-              <div style={{ flex: 1 }}>
-                <div style={{ fontSize: "0.72rem", color: "#60a5fa", fontWeight: 600, display: "flex", gap: "0.4rem", alignItems: "center" }}>
-                  Thinking
-                  {thinkingMs !== null && <span style={{ fontWeight: 400, opacity: 0.65 }}>{thinkingMs}ms</span>}
-                  {isActive && <ThinkingDots />}
-                </div>
-                <div style={{ fontSize: "0.73rem", opacity: 0.6, fontStyle: "italic", marginTop: "0.15rem" }}>
-                  {thinking.content}
-                </div>
-              </div>
-            </div>
-          )}
-
-          {/* Tool call step */}
-          {toolCall && (
-            <div style={{ display: "flex", gap: "0.6rem" }}>
-              {dot("#f59e0b")}
-              <div style={{ flex: 1 }}>
-                <div style={{ fontSize: "0.72rem", color: "#fbbf24", fontWeight: 600 }}>
-                  Tool: {toolCall.toolName}
-                </div>
-                {toolCall.params && Object.keys(toolCall.params).length > 0 ? (
-                  <div style={{ marginTop: "0.25rem", display: "flex", flexDirection: "column", gap: "0.2rem" }}>
-                    {Object.entries(toolCall.params).map(([k, v]) => (
-                      <div key={k} style={{ display: "flex", gap: "0.4rem", alignItems: "baseline", fontSize: "0.7rem" }}>
-                        <span style={{ color: "#fbbf24", opacity: 0.7, fontFamily: "monospace", flexShrink: 0 }}>{k}</span>
-                        <span style={{ color: "rgba(255,255,255,0.2)", flexShrink: 0 }}>→</span>
-                        <span style={{
-                          background: "rgba(0,0,0,0.25)", borderRadius: "4px",
-                          padding: "0.05rem 0.35rem", color: "#e5e7eb", fontFamily: "monospace",
-                          wordBreak: "break-all",
-                        }}>
-                          {typeof v === "string" ? v : JSON.stringify(v)}
-                        </span>
-                      </div>
-                    ))}
-                  </div>
-                ) : (
-                  <span style={{ fontSize: "0.7rem", color: "rgba(255,255,255,0.3)", fontStyle: "italic" }}>no params</span>
-                )}
-              </div>
-            </div>
-          )}
-
-          {/* Result step */}
-          {toolResult && (
-            <div style={{ display: "flex", gap: "0.6rem" }}>
-              {dot("#22c55e")}
-              <div style={{ flex: 1 }}>
-                {/* Result label */}
-                <div style={{ fontSize: "0.72rem", color: "#4ade80", fontWeight: 600, display: "flex", gap: "0.4rem", alignItems: "center" }}>
-                  Result
-                  {toolMs !== null && <span style={{ fontWeight: 400, opacity: 0.65 }}>{toolMs}ms</span>}
-                </div>
-                {/* Result bubble — always shows Formatted/Raw tabs */}
-                <div style={{ marginTop: "0.2rem" }}>
-                  <ChatResultBubble result={toolResult.result} />
-                </div>
-              </div>
-            </div>
-          )}
-
-          {/* Error step */}
-          {errorStep && (
-            <div style={{ display: "flex", gap: "0.6rem" }}>
-              {dot("#ef4444")}
-              <div style={{ flex: 1 }}>
-                <div style={{ fontSize: "0.72rem", color: "#fca5a5", fontWeight: 600 }}>Error</div>
-                <div style={{ fontSize: "0.73rem", color: "#fca5a5", marginTop: "0.15rem" }}>
-                  {errorStep.content}
-                </div>
-              </div>
-            </div>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
+import { ManualToolPanel } from '../components/inspector/ManualToolPanel';
+import { ChatPanel } from '../components/inspector/ChatPanel';
+import { type ChatMessage, type ExecutionRun } from '../components/inspector/chat-types';
 
 // Type for server object
 interface MCPServer {
@@ -294,26 +40,6 @@ interface LogEntry {
   message: string;
 }
 
-type ChatMessage = {
-  id: string
-  runId?: string       // groups thinking + tool_call + tool_result for one agent turn
-  type: "user" | "thinking" | "tool_call" | "tool_result" | "assistant" | "error"
-  content?: string
-  toolName?: string
-  params?: any
-  result?: any
-  resourceUri?: string // set on tool_result when tool has _meta["ui/resourceUri"]
-  timestamp: number
-  perfMark?: number    // high-res mark for duration math (performance.now())
-}
-
-type ExecutionRun = {
-  runId: string
-  serverId: string
-  startTime: number    // Date.now() — wall time
-  endTime?: number     // set on completion or error
-  steps: ChatMessage[] // thinking + tool_call + tool_result in order
-}
 interface MCPResource {
   uri: string;
   name?: string;
@@ -325,16 +51,6 @@ interface MCPResource {
 // Helper to generate unique server IDs
 const generateServerId = () => `server_${Date.now()}_${Math.random().toString(36).slice(2, 11)}`;
 
-/** Returns a compact relative time string, e.g. "2m ago", "1h ago", "just now" */
-function relativeTime(ts: number): string {
-  const secs = Math.floor((Date.now() - ts) / 1000);
-  if (secs < 10) return "just now";
-  if (secs < 60) return `${secs}s ago`;
-  const mins = Math.floor(secs / 60);
-  if (mins < 60) return `${mins}m ago`;
-  const hrs = Math.floor(mins / 60);
-  return `${hrs}h ago`;
-}
 
 export default function MCPInspector() {
 
@@ -439,45 +155,8 @@ export default function MCPInspector() {
   const [systemPromptDraft, setSystemPromptDraft] = useState("")
   const [systemPromptOpen, setSystemPromptOpen] = useState(false)
 
-  // 5A: Multi-provider LLM selector
+  // 5A: Multi-provider LLM selector (UI panel lives in 5A branch; only settings read here)
   type LLMProvider = "groq" | "openai" | "anthropic" | "gemini"
-  const PROVIDER_DEFAULTS: Record<LLMProvider, { model: string; label: string; models: { id: string; label: string }[] }> = {
-    groq: {
-      label: "Groq", model: "llama-3.1-8b-instant",
-      models: [
-        { id: "llama-3.1-8b-instant",  label: "Llama 3.1 8B (fast)" },
-        { id: "llama-3.3-70b-versatile", label: "Llama 3.3 70B" },
-        { id: "llama-3.1-70b-versatile", label: "Llama 3.1 70B" },
-        { id: "mixtral-8x7b-32768",    label: "Mixtral 8x7B" },
-      ],
-    },
-    openai: {
-      label: "OpenAI", model: "gpt-4o-mini",
-      models: [
-        { id: "gpt-4o-mini",  label: "GPT-4o Mini (fast)" },
-        { id: "gpt-4o",       label: "GPT-4o" },
-        { id: "gpt-4-turbo",  label: "GPT-4 Turbo" },
-        { id: "o1-mini",      label: "o1 Mini" },
-      ],
-    },
-    anthropic: {
-      label: "Anthropic", model: "claude-haiku-4-5-20251001",
-      models: [
-        { id: "claude-haiku-4-5-20251001",  label: "Claude Haiku 4.5 (fast)" },
-        { id: "claude-sonnet-4-5", label: "Claude Sonnet 4.5" },
-        { id: "claude-opus-4-5",   label: "Claude Opus 4.5" },
-      ],
-    },
-    gemini: {
-      label: "Gemini", model: "gemini-2.0-flash",
-      models: [
-        { id: "gemini-2.0-flash",      label: "Gemini 2.0 Flash (fast)" },
-        { id: "gemini-2.0-flash-lite", label: "Gemini 2.0 Flash Lite" },
-        { id: "gemini-1.5-pro",        label: "Gemini 1.5 Pro" },
-        { id: "gemini-1.5-flash",      label: "Gemini 1.5 Flash" },
-      ],
-    },
-  }
   const loadLLMSettings = () => {
     try {
       const raw = localStorage.getItem("fmcp_llm_settings")
@@ -485,20 +164,9 @@ export default function MCPInspector() {
     } catch { /* ignore */ }
     return { provider: "groq", model: "llama-3.1-8b-instant", apiKeys: {} }
   }
-  const [llmSettings, setLLMSettings] = useState<{
+  const [llmSettings] = useState<{
     provider: LLMProvider; model: string; apiKeys: Record<string, string>
   }>(loadLLMSettings)
-  const [llmSelectorOpen, setLLMSelectorOpen] = useState(false)
-  const [llmDraft, setLLMDraft] = useState(llmSettings)
-
-  const saveLLMSettings = (next: typeof llmSettings) => {
-    const providerChanged = next.provider !== llmSettings.provider || next.model !== llmSettings.model
-    setLLMSettings(next)
-    localStorage.setItem("fmcp_llm_settings", JSON.stringify(next))
-    if (providerChanged) {
-      setChatHistoryByServer({})
-    }
-  }
 
   // Helper to update chat for the currently selected server
   const updateChat = (updater: ChatMessage[] | ((prev: ChatMessage[]) => ChatMessage[])) => {
@@ -756,8 +424,6 @@ export default function MCPInspector() {
             : s
         )
       );
-      // Reset log offset so new logs from the restarted server aren't hidden
-      logsClearedOffsetRef.current[serverId] = 0;
       // Append reconnect notice to existing chat history (don't wipe it)
       setChatHistoryByServer(prev => ({
         ...prev,
@@ -904,20 +570,16 @@ export default function MCPInspector() {
     updateChat(prev => prev.filter((m: ChatMessage) => m.id !== thinkingMsg.id))
 
     if (res.clarification_needed) {
-      const isAuthError = res.error_type === "auth" || res.error_type === "llm"
-      // No runId — renders as a standalone red bubble below the (now-closed) run block
       const assistantMsg: ChatMessage = {
         id: generateId(),
+        runId,
         type: "assistant",
-        content: isAuthError
-          ? res.message
-          : (res.message || "Could not determine which tool to run."),
+        content: res.message,
         timestamp: Date.now(),
-        perfMark: performance.now(),
-        ...(isAuthError ? { errorType: res.error_type } : {})
+        perfMark: performance.now()
       }
       updateChat(prev => [...prev, assistantMsg])
-      // Save run with endTime so isActive becomes false in ExecutionRunBlock
+      // save run (no tool call — just clarification)
       setExecutionHistoryByServer(prev => ({
         ...prev,
         [capturedServerId]: [{ runId, serverId: capturedServerId, startTime: runStartTime, endTime: Date.now(), steps: runSteps }, ...(prev[capturedServerId] ?? [])]
@@ -975,20 +637,17 @@ export default function MCPInspector() {
 
   } catch (err: any) {
 
-    // Always clear the thinking bubble before showing the error
-    updateChat(prev => prev.filter((m: ChatMessage) => m.type !== "thinking"))
-
-    // No runId — renders as a standalone red bubble below the (now-closed) run block
     const errorMsg: ChatMessage = {
       id: generateId(),
-      type: "assistant",
-      content: err?.message || "Something went wrong. Check your LLM settings.",
+      runId,
+      type: "error",
+      content: err?.message || "Chat error",
       timestamp: Date.now(),
-      perfMark: performance.now(),
-      errorType: "llm",
-    } as any
+      perfMark: performance.now()
+    }
 
     updateChat(prev => [...prev, errorMsg])
+    runSteps.push(errorMsg)
 
     // save failed run
     setExecutionHistoryByServer(prev => ({
@@ -1217,431 +876,32 @@ export default function MCPInspector() {
                   minSize={30}
                   style={{ overflow: "hidden", display: "flex", flexDirection: "column", minHeight: 0 }}
                 >
-                  <div
-                    style={{
-                      border: "1px solid rgba(63,63,70,0.5)",
-                      borderRadius: "0.75rem",
-                      padding: "1.25rem",
-                      background: "linear-gradient(to bottom right, rgba(39,39,42,0.9), rgba(24,24,27,0.9))",
-                      flex: 1,
-                      minHeight: 0,
-                      boxSizing: "border-box",
-                      display: "flex",
-                      flexDirection: "column",
-                      overflow: "hidden",
-                    }}
-                  >
-                    {/* Header */}
-                    <div
-                      style={{
-                        display: "flex",
-                        justifyContent: "space-between",
-                        alignItems: "center",
-                        flexShrink: 0,
-                      }}
-                    >
-                      <h2 style={{ fontSize: "1.1rem", fontWeight: "600" }}>
-                        Servers
-                      </h2>
-
-                      <button
-                        style={{
-                          background: "#fff",
-                          color: "#000",
-                          border: "none",
-                          padding: "0.35rem 0.75rem",
-                          borderRadius: "0.375rem",
-                          fontSize: "0.8rem",
-                          fontWeight: "600",
-                          cursor: "pointer",
-                        }}
-                        onClick={() => setShowAddModal(true)}
-                      >
-                        + Add
-                      </button>
-                    </div>
-
-                    {/* Servers List or Empty State */}
-                    <div
-                      style={{
-                        marginTop: "1rem",
-                        padding: servers.length === 0 ? "1rem" : "0",
-                        border: servers.length === 0 ? "1px dashed rgba(63,63,70,0.6)" : "none",
-                        borderRadius: "0.5rem",
-                        textAlign: servers.length === 0 ? "center" : "left",
-                        color: "rgba(255,255,255,0.6)",
-                        flex: 1,
-                        overflow: "auto",
-                      }}
-                    >
-                      {servers.length === 0 ? (
-                        <div>No servers connected</div>
-                      ) : (
-                        servers.map((server) => {
-                          const statusConfig = {
-                            connecting: { text: "Connecting...", bg: "rgba(59,130,246,0.2)", color: "#3b82f6" },
-                            connected:  { text: "Connected",    bg: "rgba(16,185,129,0.2)",  color: "#10b981" },
-                            disconnected: { text: "Disconnected", bg: "rgba(107,114,128,0.2)", color: "#6b7280" },
-                            failed:     { text: "Failed",       bg: "rgba(239,68,68,0.2)",    color: "#ef4444" },
-                          };
-
-                          const statusInfo = statusConfig[server.status];
-                          const isSelected = selectedServer?.id === server.id;
-
-                          return (
-                            <div
-                              key={server.id}
-                              onClick={() => {
-                                if (selectedServerId === server.id) return;
-
-                                setSelectedServerId(server.id);
-                                setSelectedTool(null);
-                                setToolResult(null);
-                                setToolError(null);
-                                setToolSearch('');
-                                // 3A-3: preserve chat history on switch (no reset)
-                                
-                              }}
-                              style={{
-                                marginTop: "0.75rem",
-                                padding: "0.9rem",
-                                border: isSelected ? "1px solid rgba(99,102,241,0.5)" : "1px solid rgba(63,63,70,0.6)",
-                                borderLeft: isSelected ? "3px solid rgba(99,102,241,0.9)" : "3px solid transparent",
-                                borderRadius: "0.6rem",
-                                cursor: "pointer",
-                                background: isSelected ? "rgba(99,102,241,0.08)" : "transparent",
-                                transition: "background 0.15s, border-color 0.15s",
-                              }}
-                            >
-                              {/* HEADER */}
-                              <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center" }}>
-                                <div style={{ fontWeight: "600", fontSize: "0.9rem" }}>
-                                  {server?.name || server?.server_info?.name || "MCP Server"}
-                                </div>
-                                <span
-                                  style={{
-                                    fontSize: "0.7rem",
-                                    padding: "0.2rem 0.5rem",
-                                    borderRadius: "0.3rem",
-                                    background: statusInfo.bg,
-                                    color: statusInfo.color,
-                                    flexShrink: 0,
-                                  }}
-                                >
-                                  {statusInfo.text}
-                                </span>
-                              </div>
-
-                              {/* URL */}
-                              <div
-                                style={{
-                                  marginTop: "0.3rem",
-                                  fontSize: "0.75rem",
-                                  color: "rgba(255,255,255,0.5)",
-                                  wordBreak: "break-all",
-                                }}
-                              >
-                                {server.url}
-                              </div>
-                              <div style={{ fontSize: "0.7rem", color: "rgba(255,255,255,0.4)", display: "flex", gap: "0.6rem", flexWrap: "wrap", marginTop: "0.15rem" }}>
-                                <span>transport: {server.transport}</span>
-                                {server.status === "connected" && server.tools.length > 0 && (
-                                  <span style={{ color: "rgba(99,102,241,0.8)" }}>{server.tools.length} tool{server.tools.length !== 1 ? "s" : ""}</span>
-                                )}
-                                {server.status === "connected" && server.connectedAt && (
-                                  <span style={{ color: "rgba(34,197,94,0.6)" }}>● {relativeTime(server.connectedAt)}</span>
-                                )}
-                              </div>
-
-                              {/* ERROR */}
-                              {server.status === "failed" && server.error && (
-                                <div
-                                  style={{
-                                    marginTop: "0.5rem",
-                                    fontSize: "0.75rem",
-                                    color: "#ef4444",
-                                    padding: "0.4rem",
-                                    background: "rgba(239,68,68,0.1)",
-                                    borderRadius: "0.3rem",
-                                    wordBreak: "break-word",
-                                  }}
-                                >
-                                  {server.error}
-                                </div>
-                              )}
-
-                              {/* ACTION BUTTONS */}
-                              <div style={{ marginTop: "0.6rem", display: "flex", justifyContent: "flex-end", gap: "0.5rem" }}>
-                                {server.status === "connected" && (
-                                  <button
-                                    onClick={(e) => { e.stopPropagation(); handleDisconnect(server.id); }}
-                                    style={{
-                                      fontSize: "0.75rem", padding: "0.25rem 0.6rem",
-                                      borderRadius: "0.35rem", border: "1px solid rgba(63,63,70,0.6)",
-                                      background: "transparent", color: "#fff", cursor: "pointer",
-                                    }}
-                                  >
-                                    Disconnect
-                                  </button>
-                                )}
-
-                                {server.status === "disconnected" && (
-                                  <button
-                                    onClick={(e) => { e.stopPropagation(); handleReconnect(server.id); }}
-                                    style={{
-                                      fontSize: "0.75rem", padding: "0.25rem 0.6rem",
-                                      borderRadius: "0.35rem", border: "1px solid rgba(107,114,128,0.6)",
-                                      background: "rgba(107,114,128,0.1)", color: "#fff", cursor: "pointer",
-                                    }}
-                                  >
-                                    Reconnect
-                                  </button>
-                                )}
-
-                                {server.status === "failed" && (
-                                  <button
-                                    onClick={(e) => { e.stopPropagation(); handleReconnect(server.id); }}
-                                    style={{
-                                      fontSize: "0.75rem", padding: "0.25rem 0.6rem",
-                                      borderRadius: "0.35rem", border: "1px solid rgba(239,68,68,0.6)",
-                                      background: "rgba(239,68,68,0.1)", color: "#ef4444", cursor: "pointer",
-                                    }}
-                                  >
-                                    Retry
-                                  </button>
-                                )}
-
-                                {server.status !== "connecting" && (
-                                  <button
-                                    onClick={(e) => { e.stopPropagation(); handleRemove(server.id); }}
-                                    style={{
-                                      fontSize: "0.75rem", padding: "0.25rem 0.4rem",
-                                      borderRadius: "0.35rem", border: "1px solid rgba(63,63,70,0.6)",
-                                      background: "transparent", color: "rgba(255,255,255,0.6)", cursor: "pointer",
-                                    }}
-                                    title="Remove server"
-                                  >
-                                    ✕
-                                  </button>
-                                )}
-                              </div>
-
-                              {/* TOOLS (only show when selected) */}
-                              {isSelected && (server?.tools?.length ?? 0) > 0 && (
-                                <>
-                                  {/* 5C: Tools | Saved sub-tabs */}
-                                  <div style={{ display: "flex", margin: "0.4rem 0.5rem 0.2rem", gap: "0.25rem" }} onClick={e => e.stopPropagation()}>
-                                    {(["tools", "saved"] as const).map(tab => (
-                                      <button
-                                        key={tab}
-                                        onClick={() => setToolSubTab(tab)}
-                                        style={{
-                                          flex: 1, fontSize: "0.68rem", padding: "0.18rem 0",
-                                          borderRadius: "4px", cursor: "pointer",
-                                          border: toolSubTab === tab ? "1px solid rgba(99,102,241,0.5)" : "1px solid rgba(63,63,70,0.4)",
-                                          background: toolSubTab === tab ? "rgba(99,102,241,0.15)" : "transparent",
-                                          color: toolSubTab === tab ? "rgba(165,180,252,0.9)" : "rgba(255,255,255,0.4)",
-                                          fontWeight: toolSubTab === tab ? 600 : 400,
-                                        }}
-                                      >
-                                        {tab === "tools" ? `Tools (${server.tools.length})` : `Saved (${savedRequests.length})`}
-                                      </button>
-                                    ))}
-                                  </div>
-
-                                  {toolSubTab === "tools" && (
-                                    <div style={{ margin: "0 0.5rem 0.2rem", position: "relative" }}>
-                                      <input
-                                        value={toolSearch}
-                                        onChange={e => setToolSearch(e.target.value)}
-                                        placeholder="Search tools..."
-                                        onClick={e => e.stopPropagation()}
-                                        style={{
-                                          width: "100%", boxSizing: "border-box",
-                                          padding: "0.25rem 1.4rem 0.25rem 1.5rem",
-                                          fontSize: "0.7rem", borderRadius: "4px",
-                                          border: "1px solid rgba(63,63,70,0.5)",
-                                          background: "rgba(0,0,0,0.25)", color: "#fff",
-                                        }}
-                                      />
-                                      <span style={{ position: "absolute", left: "0.4rem", top: "50%", transform: "translateY(-50%)", fontSize: "0.65rem", color: "rgba(255,255,255,0.3)", pointerEvents: "none" }}>⌕</span>
-                                      {toolSearch && (
-                                        <span
-                                          onClick={e => { e.stopPropagation(); setToolSearch(''); }}
-                                          style={{ position: "absolute", right: "0.4rem", top: "50%", transform: "translateY(-50%)", fontSize: "0.65rem", color: "rgba(255,255,255,0.4)", cursor: "pointer", lineHeight: 1 }}
-                                        >✕</span>
-                                      )}
-                                    </div>
-                                  )}
-
-                                  {/* 5C: Saved requests list */}
-                                  {toolSubTab === "saved" && (
-                                    <div onClick={e => e.stopPropagation()}>
-                                      {savedRequests.length === 0 ? (
-                                        <div style={{ margin: "0.5rem", fontSize: "0.72rem", color: "rgba(255,255,255,0.3)", textAlign: "center", padding: "0.5rem" }}>
-                                          No saved requests yet.<br />Run a tool and click "Save".
-                                        </div>
-                                      ) : savedRequests.map(req => (
-                                        <div
-                                          key={req.id}
-                                          style={{
-                                            margin: "0.3rem 0.5rem",
-                                            padding: "0.35rem 0.5rem",
-                                            borderRadius: "5px",
-                                            border: "1px solid rgba(63,63,70,0.4)",
-                                            background: "rgba(0,0,0,0.2)",
-                                            cursor: renamingId === req.id ? "default" : "pointer",
-                                          }}
-                                          onClick={() => {
-                                            if (renamingId === req.id) return;
-                                            const tool = server.tools.find((t: any) => t.name === req.toolName);
-                                            if (tool) {
-                                              setSelectedTool(tool);
-                                              setFormPrefill(req.params as Record<string, any>);
-                                              setToolResult(null);
-                                              setToolError(null);
-                                              setToolSubTab("tools");
-                                            }
-                                          }}
-                                        >
-                                          {renamingId === req.id ? (
-                                            <div style={{ display: "flex", gap: "0.3rem", alignItems: "center" }} onClick={e => e.stopPropagation()}>
-                                              <input
-                                                autoFocus
-                                                value={renameTitle}
-                                                onChange={e => setRenameTitle(e.target.value)}
-                                                onKeyDown={e => {
-                                                  if (e.key === "Enter" && renameTitle.trim()) {
-                                                    renameSavedRequest(selectedServer.url, req.id, renameTitle.trim());
-                                                    setSavedRequests(loadSavedRequests(selectedServer.url));
-                                                    setRenamingId(null);
-                                                  } else if (e.key === "Escape") {
-                                                    setRenamingId(null);
-                                                  }
-                                                }}
-                                                style={{
-                                                  flex: 1, fontSize: "0.72rem", padding: "0.15rem 0.3rem",
-                                                  borderRadius: "3px", border: "1px solid rgba(99,102,241,0.5)",
-                                                  background: "rgba(0,0,0,0.35)", color: "#fff",
-                                                }}
-                                              />
-                                              <button
-                                                onClick={() => {
-                                                  if (renameTitle.trim()) {
-                                                    renameSavedRequest(selectedServer.url, req.id, renameTitle.trim());
-                                                    setSavedRequests(loadSavedRequests(selectedServer.url));
-                                                  }
-                                                  setRenamingId(null);
-                                                }}
-                                                style={{
-                                                  fontSize: "0.6rem", padding: "0.1rem 0.3rem",
-                                                  borderRadius: "3px", border: "1px solid rgba(99,102,241,0.4)",
-                                                  background: "rgba(99,102,241,0.15)", color: "rgba(165,180,252,0.9)", cursor: "pointer", flexShrink: 0,
-                                                }}
-                                              >✓</button>
-                                              <button
-                                                onClick={() => setRenamingId(null)}
-                                                style={{
-                                                  fontSize: "0.6rem", padding: "0.1rem 0.3rem",
-                                                  borderRadius: "3px", border: "1px solid rgba(63,63,70,0.5)",
-                                                  background: "transparent", color: "rgba(255,255,255,0.3)", cursor: "pointer", flexShrink: 0,
-                                                }}
-                                              >✕</button>
-                                            </div>
-                                          ) : (
-                                            <div style={{ display: "flex", justifyContent: "space-between", alignItems: "flex-start" }}>
-                                              <div style={{ flex: 1, minWidth: 0 }}>
-                                                <div style={{ fontSize: "0.72rem", color: "#fff", fontWeight: 500, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>{req.title}</div>
-                                                <div style={{ fontSize: "0.64rem", color: "rgba(255,255,255,0.4)", marginTop: "0.1rem" }}>{req.toolName}</div>
-                                              </div>
-                                              <div style={{ display: "flex", gap: "0.2rem", flexShrink: 0, marginLeft: "0.35rem" }}>
-                                                <button
-                                                  onClick={e => {
-                                                    e.stopPropagation();
-                                                    setRenamingId(req.id);
-                                                    setRenameTitle(req.title);
-                                                  }}
-                                                  style={{
-                                                    fontSize: "0.6rem", padding: "0.1rem 0.3rem",
-                                                    borderRadius: "3px", border: "1px solid rgba(63,63,70,0.5)",
-                                                    background: "transparent", color: "rgba(255,255,255,0.3)", cursor: "pointer",
-                                                  }}
-                                                  title="Rename saved request"
-                                                >✎</button>
-                                                <button
-                                                  onClick={e => {
-                                                    e.stopPropagation();
-                                                    deleteSavedRequest(selectedServer.url, req.id);
-                                                    setSavedRequests(loadSavedRequests(selectedServer.url));
-                                                  }}
-                                                  style={{
-                                                    fontSize: "0.6rem", padding: "0.1rem 0.3rem",
-                                                    borderRadius: "3px", border: "1px solid rgba(63,63,70,0.5)",
-                                                    background: "transparent", color: "rgba(255,255,255,0.3)", cursor: "pointer",
-                                                  }}
-                                                  title="Delete saved request"
-                                                >✕</button>
-                                              </div>
-                                            </div>
-                                          )}
-                                        </div>
-                                      ))}
-                                    </div>
-                                  )}
-                                </>
-                              )}
-                              {isSelected && toolSubTab === "tools" && server?.tools?.filter((t: any) =>
-                                toolSearch.trim() === '' || t.name.toLowerCase().includes(toolSearch.toLowerCase()) || (t.description ?? '').toLowerCase().includes(toolSearch.toLowerCase())
-                              ).map((tool: any) => {
-                                const ann = tool.annotations ?? {};
-                                const badges = [
-                                  ann.readOnlyHint    && { label: "Read-only",   bg: "rgba(63,63,70,0.5)",    color: "rgba(255,255,255,0.5)" },
-                                  ann.destructiveHint && { label: "Destructive", bg: "rgba(239,68,68,0.15)",  color: "rgba(252,165,165,0.9)" },
-                                  ann.idempotentHint  && { label: "Idempotent",  bg: "rgba(34,197,94,0.12)",  color: "rgba(134,239,172,0.85)" },
-                                  ann.openWorldHint   && { label: "External",    bg: "rgba(59,130,246,0.15)", color: "rgba(147,197,253,0.9)" },
-                                ].filter(Boolean) as { label: string; bg: string; color: string }[];
-                                return (
-                                  <div
-                                    key={tool.name}
-                                    onClick={(e) => {
-                                      e.stopPropagation();
-                                      setSelectedTool(tool);
-                                      setFormPrefill(undefined);
-                                      setToolResult(null);
-                                      setToolError(null);
-                                    }}
-                                    style={{
-                                      marginLeft: "0.5rem",
-                                      marginTop: "0.4rem",
-                                      fontSize: "0.82rem",
-                                      cursor: "pointer",
-                                      color: selectedTool?.name === tool.name ? "#fff" : "rgba(255,255,255,0.65)",
-                                      padding: "0.2rem 0.4rem",
-                                      borderRadius: "0.25rem",
-                                      background: selectedTool?.name === tool.name ? "rgba(255,255,255,0.08)" : "transparent",
-                                    }}
-                                  >
-                                    <div>• {tool.name}</div>
-                                    {badges.length > 0 && (
-                                      <div style={{ display: "flex", gap: "0.25rem", flexWrap: "wrap", marginTop: "0.2rem", marginLeft: "0.6rem" }}>
-                                        {badges.map(b => (
-                                          <span key={b.label} style={{
-                                            fontSize: "0.6rem", padding: "0.05rem 0.3rem", borderRadius: "999px",
-                                            background: b.bg, color: b.color, fontWeight: 500,
-                                          }}>{b.label}</span>
-                                        ))}
-                                      </div>
-                                    )}
-                                  </div>
-                                );
-                              })}
-                            </div>
-                          );
-                        })
-                      )}
-                    </div>
-
-                  </div>
+                  <ServerListPanel
+                    servers={servers}
+                    selectedServerId={selectedServerId}
+                    selectedServer={selectedServer}
+                    selectedTool={selectedTool}
+                    toolSubTab={toolSubTab}
+                    toolSearch={toolSearch}
+                    savedRequests={savedRequests}
+                    renamingId={renamingId}
+                    renameTitle={renameTitle}
+                    setSelectedServerId={setSelectedServerId}
+                    setSelectedTool={setSelectedTool}
+                    setToolResult={setToolResult}
+                    setToolError={setToolError}
+                    setToolSearch={setToolSearch}
+                    setToolSubTab={setToolSubTab}
+                    setSavedRequests={setSavedRequests}
+                    setFormPrefill={setFormPrefill}
+                    setRenamingId={setRenamingId}
+                    setRenameTitle={setRenameTitle}
+                    renameSavedRequest={renameSavedRequest}
+                    onDisconnect={handleDisconnect}
+                    onReconnect={handleReconnect}
+                    onRemove={handleRemove}
+                    onAddServer={() => setShowAddModal(true)}
+                  />
                 </Panel>
 
                 {/* ── Vertical resize handle ─────────────────────────────── */}
@@ -1816,589 +1076,48 @@ export default function MCPInspector() {
                 {/* ── MANUAL MODE ─── */}
                 <div style={{ flex: 1, display: "flex", flexDirection: "column", overflow: "hidden", minHeight: 0 }}>
                   {mode === "manual" && selectedServer && selectedTool && (
-                    <div style={{ overflowY: "auto", flex: 1, minHeight: 0, paddingBottom: "0.5rem", marginTop: "1rem" }}>
-                      <div style={{ display: "flex", alignItems: "center", gap: "0.6rem", flexWrap: "wrap", marginBottom: "1rem" }}>
-                        <h3 style={{ margin: 0 }}>{selectedTool.name}</h3>
-                        {(() => {
-                          const ann = selectedTool.annotations ?? {};
-                          const ANNOTATION_META: { key: string; label: string; bg: string; color: string; tip: string }[] = [
-                            { key: "readOnlyHint",    label: "Read-only",   bg: "rgba(63,63,70,0.6)",    color: "rgba(220,220,220,0.85)", tip: "Tool does not modify state" },
-                            { key: "destructiveHint", label: "Destructive", bg: "rgba(239,68,68,0.18)",  color: "rgba(252,165,165,0.95)", tip: "Tool may delete or overwrite data" },
-                            { key: "idempotentHint",  label: "Idempotent",  bg: "rgba(34,197,94,0.15)",  color: "rgba(134,239,172,0.9)",  tip: "Same call with same args always gives same result" },
-                            { key: "openWorldHint",   label: "External",    bg: "rgba(59,130,246,0.18)", color: "rgba(147,197,253,0.95)", tip: "Tool interacts with the external world (web, APIs, filesystem)" },
-                          ];
-                          return ANNOTATION_META.filter(a => ann[a.key]).map(a => (
-                            <span key={a.label} title={a.tip} style={{
-                              fontSize: "0.65rem", padding: "0.15rem 0.55rem", borderRadius: "999px",
-                              background: a.bg, color: a.color, fontWeight: 600, cursor: "help",
-                              border: `1px solid ${a.color}33`,
-                            }}>{a.label}</span>
-                          ));
-                        })()}
-                      </div>
-
-                      {selectedServer.status === "connected" ? (
-                        <>
-                          <JsonSchemaForm
-                            schema={selectedTool.inputSchema}
-                            initialValues={formPrefill}
-                            onSubmit={runTool}
-                            submitLabel="Run Tool"
-                            loading={executing}
-                          />
-
-                          {lastRunParams && (
-                            <div style={{ marginTop: "0.75rem", display: "flex", alignItems: "center", gap: "0.5rem", flexWrap: "wrap" }}>
-                              <button
-                                onClick={() => {
-                                  const jsonrpc = {
-                                    jsonrpc: "2.0", id: 1,
-                                    method: "tools/call",
-                                    params: { name: selectedTool.name, arguments: lastRunParams },
-                                  };
-                                  navigator.clipboard.writeText(JSON.stringify(jsonrpc, null, 2));
-                                  setCopyRequestToast(true);
-                                  setTimeout(() => setCopyRequestToast(false), 2000);
-                                }}
-                                style={{
-                                  fontSize: "0.72rem", padding: "0.25rem 0.65rem", borderRadius: "6px",
-                                  background: "rgba(99,102,241,0.1)", border: "1px solid rgba(99,102,241,0.35)",
-                                  color: "rgba(165,180,252,0.9)", cursor: "pointer",
-                                }}
-                              >
-                                {copyRequestToast ? "✓ Copied!" : "Copy as JSON-RPC"}
-                              </button>
-                              {/* 5C: Save this request */}
-                              <button
-                                onClick={() => {
-                                  setSaveTitle(selectedTool.name);
-                                  setSaveDialogOpen(true);
-                                }}
-                                style={{
-                                  fontSize: "0.72rem", padding: "0.25rem 0.65rem", borderRadius: "6px",
-                                  background: "rgba(34,197,94,0.08)", border: "1px solid rgba(34,197,94,0.3)",
-                                  color: "rgba(134,239,172,0.9)", cursor: "pointer",
-                                }}
-                              >
-                                Save Request
-                              </button>
-                            </div>
-                          )}
-
-                          {/* 5C: Save dialog */}
-                          {saveDialogOpen && (
-                            <div style={{
-                              position: "fixed", inset: 0, zIndex: 9999,
-                              background: "rgba(0,0,0,0.6)", display: "flex", alignItems: "center", justifyContent: "center",
-                            }} onClick={() => setSaveDialogOpen(false)}>
-                              <div
-                                onClick={e => e.stopPropagation()}
-                                style={{
-                                  background: "#18181b", border: "1px solid rgba(63,63,70,0.7)",
-                                  borderRadius: "10px", padding: "1.25rem", width: "320px",
-                                }}
-                              >
-                                <div style={{ fontSize: "0.9rem", fontWeight: 600, color: "#fff", marginBottom: "0.75rem" }}>Save Request</div>
-                                <input
-                                  autoFocus
-                                  value={saveTitle}
-                                  onChange={e => setSaveTitle(e.target.value)}
-                                  onKeyDown={e => {
-                                    if (e.key === "Enter" && saveTitle.trim() && lastRunParams && selectedServer) {
-                                      saveRequest(selectedServer.url, {
-                                        id: crypto.randomUUID(),
-                                        title: saveTitle.trim(),
-                                        toolName: selectedTool.name,
-                                        params: lastRunParams,
-                                        createdAt: Date.now(),
-                                        serverUrl: selectedServer.url,
-                                      });
-                                      setSavedRequests(loadSavedRequests(selectedServer.url));
-                                      setSaveDialogOpen(false);
-                                    }
-                                    if (e.key === "Escape") setSaveDialogOpen(false);
-                                  }}
-                                  placeholder="Request title..."
-                                  style={{
-                                    width: "100%", boxSizing: "border-box",
-                                    padding: "0.45rem 0.6rem", fontSize: "0.82rem",
-                                    borderRadius: "6px", border: "1px solid rgba(63,63,70,0.6)",
-                                    background: "rgba(0,0,0,0.3)", color: "#fff", outline: "none",
-                                    marginBottom: "0.75rem",
-                                  }}
-                                />
-                                <div style={{ display: "flex", gap: "0.5rem", justifyContent: "flex-end" }}>
-                                  <button
-                                    onClick={() => setSaveDialogOpen(false)}
-                                    style={{
-                                      fontSize: "0.78rem", padding: "0.3rem 0.75rem", borderRadius: "6px",
-                                      border: "1px solid rgba(63,63,70,0.6)", background: "transparent",
-                                      color: "rgba(255,255,255,0.5)", cursor: "pointer",
-                                    }}
-                                  >Cancel</button>
-                                  <button
-                                    onClick={() => {
-                                      if (!saveTitle.trim() || !lastRunParams || !selectedServer) return;
-                                      saveRequest(selectedServer.url, {
-                                        id: crypto.randomUUID(),
-                                        title: saveTitle.trim(),
-                                        toolName: selectedTool.name,
-                                        params: lastRunParams,
-                                        createdAt: Date.now(),
-                                        serverUrl: selectedServer.url,
-                                      });
-                                      setSavedRequests(loadSavedRequests(selectedServer.url));
-                                      setSaveDialogOpen(false);
-                                    }}
-                                    disabled={!saveTitle.trim()}
-                                    style={{
-                                      fontSize: "0.78rem", padding: "0.3rem 0.75rem", borderRadius: "6px",
-                                      border: "1px solid rgba(34,197,94,0.4)", background: "rgba(34,197,94,0.1)",
-                                      color: "rgba(134,239,172,0.9)", cursor: saveTitle.trim() ? "pointer" : "not-allowed",
-                                      opacity: saveTitle.trim() ? 1 : 0.5,
-                                    }}
-                                  >Save</button>
-                                </div>
-                              </div>
-                            </div>
-                          )}
-
-                          {(toolResult || toolError) && (
-                            <div style={{ marginTop: "1.25rem" }}>
-                              <ToolResult
-                                result={toolResult}
-                                error={toolError || undefined}
-                                executionTime={executionTime}
-                              />
-                            </div>
-                          )}
-                        </>
-                      ) : (
-                        <div style={{
-                          padding: "1rem", border: "1px dashed rgba(239,68,68,0.6)",
-                          borderRadius: "0.5rem", textAlign: "center", color: "rgba(239,68,68,0.8)",
-                        }}>
-                          Server is not connected. Please reconnect to run tools.
-                        </div>
-                      )}
-                    </div>
+                    <ManualToolPanel
+                      selectedTool={selectedTool}
+                      selectedServer={selectedServer}
+                      formPrefill={formPrefill}
+                      executing={executing}
+                      toolResult={toolResult}
+                      toolError={toolError}
+                      executionTime={executionTime}
+                      lastRunParams={lastRunParams}
+                      copyRequestToast={copyRequestToast}
+                      saveDialogOpen={saveDialogOpen}
+                      saveTitle={saveTitle}
+                      setSaveTitle={setSaveTitle}
+                      setSaveDialogOpen={setSaveDialogOpen}
+                      setCopyRequestToast={setCopyRequestToast}
+                      setSavedRequests={setSavedRequests}
+                      onRunTool={runTool}
+                    />
                   )}
 
                   {/* ── CHAT MODE ─── */}
                   {mode === "chat" && selectedServer && (
-                    <div style={{ display: "flex", flexDirection: "column", flex: 1, minHeight: 0, overflow: "hidden" }}>
-                      {selectedServer.status === "connected" ? (
-                        <>
-                          {/* Chat History — fills remaining flex space, scrolls */}
-                          <div
-                            ref={chatRef}
-                            style={{
-                              flex: 1,
-                              minHeight: 0,
-                              overflowY: "auto",
-                              padding: "0.5rem 0.5rem 2rem 0.25rem"
-                            }}
-                          >
-                            {/* Inner wrapper: flex layout for gap spacing.
-                                Kept separate from the scroll container so flex items
-                                don't shrink-to-fit and prevent overflow scrolling. */}
-                            <div style={{ display: "flex", flexDirection: "column", gap: "0.5rem" }}>
-                            {groupMessages(chatHistory, executionHistory).map((group, _idx, _arr) => {
-                              const isLast = _idx === _arr.length - 1;
-                              if (group.kind === "run") {
-                                const toolResult = group.steps.find(s => s.type === "tool_result");
-                                const toolCall = group.steps.find(s => s.type === "tool_call");
-                                const sessionId = selectedServer?.session_id ?? undefined;
-                                const hasWidget = !!(toolResult?.resourceUri && sessionId && toolCall);
-                                return (
-                                  <React.Fragment key={group.runId}>
-                                    <ExecutionRunBlock steps={group.steps} run={group.run} />
-                                    {hasWidget && (
-                                      <div style={{
-                                        width: "100%",
-                                        border: "1px solid rgba(99,102,241,0.3)",
-                                        borderRadius: "10px",
-                                        overflow: "hidden",
-                                        background: "rgba(99,102,241,0.04)",
-                                      }}>
-                                        <WidgetSandbox
-                                          sessionId={sessionId!}
-                                          resourceUri={toolResult!.resourceUri!}
-                                          toolInput={toolCall!.params || {}}
-                                          toolResult={toolResult!.result}
-                                        />
-                                      </div>
-                                    )}
-                                    {/* Scroll anchor: AFTER widget so auto-scroll reveals the full widget */}
-                                    {isLast && <div ref={chatBottomRef} />}
-                                  </React.Fragment>
-                                );
-                              }
-
-                              const msg = group.msg;
-
-                              if (msg.type === "user") {
-                                return (
-                                  <React.Fragment key={msg.id}>
-                                    <div style={{ display: "flex", justifyContent: "flex-end" }}>
-                                      <div style={{
-                                        background: "rgba(37,99,235,0.85)",
-                                        border: "1px solid rgba(59,130,246,0.4)",
-                                        color: "#fff",
-                                        padding: "0.55rem 0.85rem",
-                                        borderRadius: "16px 16px 4px 16px",
-                                        maxWidth: "75%", fontSize: "0.875rem", lineHeight: 1.5,
-                                      }}>
-                                        {msg.content}
-                                      </div>
-                                    </div>
-                                    {isLast && <div ref={chatBottomRef} />}
-                                  </React.Fragment>
-                                );
-                              }
-
-                              if (msg.type === "assistant") {
-                                const isErrMsg = !!(msg as any).errorType
-                                return (
-                                  <React.Fragment key={msg.id}>
-                                    <div style={{ display: "flex", justifyContent: "flex-start", gap: "0.5rem", alignItems: "flex-start" }}>
-                                      <div style={{
-                                        width: "22px", height: "22px", borderRadius: "50%", flexShrink: 0, marginTop: "2px",
-                                        background: isErrMsg ? "rgba(239,68,68,0.15)" : "rgba(99,102,241,0.2)",
-                                        border: `1px solid ${isErrMsg ? "rgba(239,68,68,0.4)" : "rgba(99,102,241,0.4)"}`,
-                                        display: "flex", alignItems: "center", justifyContent: "center",
-                                        fontSize: "0.65rem",
-                                      }}>{isErrMsg ? "⚠" : "🤖"}</div>
-                                      <div style={{
-                                        background: isErrMsg ? "rgba(239,68,68,0.08)" : "rgba(39,39,42,0.8)",
-                                        border: `1px solid ${isErrMsg ? "rgba(239,68,68,0.35)" : "rgba(63,63,70,0.6)"}`,
-                                        padding: "0.55rem 0.85rem",
-                                        borderRadius: "4px 16px 16px 16px",
-                                        maxWidth: "75%", fontSize: "0.875rem", lineHeight: 1.5,
-                                        color: isErrMsg ? "rgba(252,165,165,0.9)" : "rgba(255,255,255,0.85)",
-                                        display: "flex", flexDirection: "column", gap: "0.5rem",
-                                      }}>
-                                        <span>{msg.content}</span>
-                                        {isErrMsg && (
-                                          <button
-                                            onClick={() => { setLLMDraft(llmSettings); setLLMSelectorOpen(true); }}
-                                            style={{
-                                              fontSize: "0.7rem", padding: "0.25rem 0.6rem", borderRadius: "6px", cursor: "pointer",
-                                              background: "rgba(239,68,68,0.15)", border: "1px solid rgba(239,68,68,0.35)",
-                                              color: "rgba(252,165,165,0.9)", alignSelf: "flex-start",
-                                            }}
-                                          >
-                                            Fix LLM settings
-                                          </button>
-                                        )}
-                                      </div>
-                                    </div>
-                                    {isLast && <div ref={chatBottomRef} />}
-                                  </React.Fragment>
-                                );
-                              }
-
-                              return null;
-                            })}
-                            </div>{/* close inner flex wrapper */}
-                          </div>
-
-                          {/* Starter prompt chips — shown when no user message sent yet */}
-                          {!chatHistory.some(m => m.type === "user") && (selectedServer?.tools?.length ?? 0) > 0 && (
-                            <div style={{ flexShrink: 0, padding: "0.5rem 0.25rem", display: "flex", flexWrap: "wrap", gap: "0.4rem" }}>
-                              {[
-                                ...selectedServer!.tools.slice(0, 3).map((t: any) => `Try ${t.name} with example data`),
-                                selectedServer!.tools.length > 1 ? `Run ${selectedServer!.tools[0].name}` : null,
-                              ].filter(Boolean).slice(0, 4).map((prompt: any) => (
-                                <button
-                                  key={prompt}
-                                  onClick={() => runChatToolWithMessage(prompt)}
-                                  style={{
-                                    fontSize: "0.72rem", padding: "0.3rem 0.7rem", borderRadius: "999px",
-                                    background: "rgba(99,102,241,0.08)", border: "1px solid rgba(99,102,241,0.25)",
-                                    color: "rgba(165,180,252,0.85)", cursor: "pointer",
-                                    transition: "background 0.15s",
-                                  }}
-                                  onMouseEnter={e => (e.currentTarget.style.background = "rgba(99,102,241,0.18)")}
-                                  onMouseLeave={e => (e.currentTarget.style.background = "rgba(99,102,241,0.08)")}
-                                >
-                                  {prompt}
-                                </button>
-                              ))}
-                            </div>
-                          )}
-
-                          {/* Chat Input — always at bottom, shrinks to its content */}
-                          <div style={{ flexShrink: 0, display: "flex", flexDirection: "column", gap: "0.35rem", paddingTop: "0.5rem", borderTop: "1px solid rgba(63,63,70,0.3)" }}>
-
-                            {/* ── Toolbar row: model badge · system prompt · clear ── */}
-                            <div style={{ display: "flex", alignItems: "center", gap: "0.4rem", paddingTop: "0.3rem" }}>
-                              {/* Model badge — clickable to open LLM selector */}
-                              <button
-                                onClick={() => { setLLMDraft(llmSettings); setLLMSelectorOpen(true); }}
-                                title="Change LLM provider / model"
-                                style={{
-                                  fontSize: "0.65rem", padding: "0.15rem 0.5rem", borderRadius: "999px",
-                                  background: "rgba(99,102,241,0.12)", border: "1px solid rgba(99,102,241,0.3)",
-                                  color: "rgba(165,180,252,0.9)", fontFamily: "monospace", flexShrink: 0,
-                                  cursor: "pointer",
-                                }}
-                              >
-                                {PROVIDER_DEFAULTS[llmSettings.provider]?.label ?? llmSettings.provider} · {llmSettings.model.length > 20 ? llmSettings.model.slice(0, 20) + "…" : llmSettings.model}
-                              </button>
-
-                              {/* System prompt button */}
-                              <button
-                                onClick={() => { setSystemPromptDraft(systemPrompt); setSystemPromptOpen(true); }}
-                                title="System prompt"
-                                style={{
-                                  fontSize: "0.65rem", padding: "0.15rem 0.5rem", borderRadius: "999px",
-                                  background: systemPrompt ? "rgba(99,102,241,0.15)" : "transparent",
-                                  border: systemPrompt ? "1px solid rgba(99,102,241,0.4)" : "1px solid rgba(63,63,70,0.5)",
-                                  color: systemPrompt ? "rgba(165,180,252,0.9)" : "rgba(255,255,255,0.35)",
-                                  cursor: "pointer", flexShrink: 0,
-                                }}
-                              >
-                                ⚙ {systemPrompt ? "Prompt set" : "System prompt"}
-                              </button>
-
-                              {/* Spacer */}
-                              <div style={{ flex: 1 }} />
-
-                              {/* Clear chat */}
-                              {chatHistory.length > 1 && selectedServerId && (
-                                <button
-                                  onClick={() => setChatHistoryByServer(prev => ({ ...prev, [selectedServerId]: [] }))}
-                                  style={{ fontSize: "0.65rem", background: "none", border: "1px solid rgba(63,63,70,0.5)", borderRadius: "4px", color: "rgba(255,255,255,0.35)", cursor: "pointer", padding: "0.1rem 0.4rem", flexShrink: 0 }}
-                                >
-                                  Clear Chat
-                                </button>
-                              )}
-                            </div>
-
-                            {/* ── System prompt dialog overlay ── */}
-                            {systemPromptOpen && (
-                              <div style={{
-                                position: "fixed", inset: 0, zIndex: 50,
-                                background: "rgba(0,0,0,0.6)", display: "flex", alignItems: "center", justifyContent: "center",
-                              }}
-                                onClick={(e) => { if (e.target === e.currentTarget) setSystemPromptOpen(false); }}
-                              >
-                                <div style={{
-                                  background: "#18181b", border: "1px solid rgba(63,63,70,0.7)",
-                                  borderRadius: "12px", padding: "1.25rem", width: "420px", maxWidth: "90vw",
-                                  display: "flex", flexDirection: "column", gap: "0.75rem",
-                                }}>
-                                  <div style={{ fontSize: "0.9rem", fontWeight: 600, color: "rgba(255,255,255,0.85)" }}>System Prompt</div>
-                                  <textarea
-                                    value={systemPromptDraft}
-                                    onChange={e => setSystemPromptDraft(e.target.value)}
-                                    placeholder="You are a helpful assistant with access to MCP tools."
-                                    rows={5}
-                                    style={{
-                                      background: "#09090b", border: "1px solid rgba(63,63,70,0.6)",
-                                      borderRadius: "6px", padding: "0.5rem 0.6rem",
-                                      color: "#fff", fontSize: "0.8rem", resize: "vertical",
-                                      fontFamily: "inherit", lineHeight: 1.5,
-                                    }}
-                                  />
-                                  <div style={{ display: "flex", justifyContent: "flex-end", gap: "0.5rem" }}>
-                                    <button
-                                      onClick={() => setSystemPromptOpen(false)}
-                                      style={{ fontSize: "0.75rem", padding: "0.35rem 0.75rem", borderRadius: "6px", background: "transparent", border: "1px solid rgba(63,63,70,0.5)", color: "rgba(255,255,255,0.5)", cursor: "pointer" }}
-                                    >
-                                      Cancel
-                                    </button>
-                                    <button
-                                      onClick={() => { setSystemPrompt(systemPromptDraft); setSystemPromptOpen(false); }}
-                                      style={{ fontSize: "0.75rem", padding: "0.35rem 0.75rem", borderRadius: "6px", background: "rgba(99,102,241,0.8)", border: "none", color: "#fff", cursor: "pointer", fontWeight: 600 }}
-                                    >
-                                      Save
-                                    </button>
-                                  </div>
-                                </div>
-                              </div>
-                            )}
-
-                            {/* ── LLM selector dialog ── */}
-                            {llmSelectorOpen && (
-                              <div style={{
-                                position: "fixed", inset: 0, zIndex: 50,
-                                background: "rgba(0,0,0,0.6)", display: "flex", alignItems: "center", justifyContent: "center",
-                              }}
-                                onClick={(e) => { if (e.target === e.currentTarget) setLLMSelectorOpen(false); }}
-                              >
-                                <div style={{
-                                  background: "#18181b", border: "1px solid rgba(63,63,70,0.7)",
-                                  borderRadius: "12px", padding: "1.25rem", width: "420px", maxWidth: "90vw",
-                                  display: "flex", flexDirection: "column", gap: "0.85rem",
-                                }}>
-                                  <div style={{ fontSize: "0.9rem", fontWeight: 600, color: "rgba(255,255,255,0.85)" }}>LLM Settings</div>
-
-                                  {/* Warning when provider/model differs from saved */}
-                                  {(llmDraft.provider !== llmSettings.provider || llmDraft.model !== llmSettings.model) && chatHistory.length > 0 && (
-                                    <div style={{
-                                      fontSize: "0.72rem", padding: "0.4rem 0.65rem", borderRadius: "6px",
-                                      background: "rgba(234,179,8,0.1)", border: "1px solid rgba(234,179,8,0.3)",
-                                      color: "rgba(253,224,71,0.85)", display: "flex", alignItems: "center", gap: "0.4rem",
-                                    }}>
-                                      ⚠ Changing provider or model will clear the current chat
-                                    </div>
-                                  )}
-
-                                  {/* Provider selector */}
-                                  <div style={{ display: "flex", flexDirection: "column", gap: "0.35rem" }}>
-                                    <label style={{ fontSize: "0.75rem", color: "rgba(255,255,255,0.5)" }}>Provider</label>
-                                    <div style={{ display: "flex", gap: "0.4rem", flexWrap: "wrap" }}>
-                                      {(["groq", "openai", "anthropic", "gemini"] as LLMProvider[]).map(p => (
-                                        <button key={p}
-                                          onClick={() => setLLMDraft(d => ({
-                                            ...d,
-                                            provider: p,
-                                            model: PROVIDER_DEFAULTS[p].model,
-                                          }))}
-                                          style={{
-                                            fontSize: "0.75rem", padding: "0.3rem 0.7rem", borderRadius: "6px", cursor: "pointer",
-                                            background: llmDraft.provider === p ? "rgba(99,102,241,0.8)" : "rgba(39,39,42,0.8)",
-                                            border: llmDraft.provider === p ? "1px solid rgba(99,102,241,0.6)" : "1px solid rgba(63,63,70,0.5)",
-                                            color: llmDraft.provider === p ? "#fff" : "rgba(255,255,255,0.6)",
-                                            fontWeight: llmDraft.provider === p ? 600 : 400,
-                                          }}
-                                        >
-                                          {PROVIDER_DEFAULTS[p].label}
-                                        </button>
-                                      ))}
-                                    </div>
-                                  </div>
-
-                                  {/* Model selector */}
-                                  <div style={{ display: "flex", flexDirection: "column", gap: "0.35rem" }}>
-                                    <label style={{ fontSize: "0.75rem", color: "rgba(255,255,255,0.5)" }}>Model</label>
-                                    <select
-                                      value={PROVIDER_DEFAULTS[llmDraft.provider].models.some(m => m.id === llmDraft.model) ? llmDraft.model : "__custom__"}
-                                      onChange={e => {
-                                        if (e.target.value !== "__custom__") {
-                                          setLLMDraft(d => ({ ...d, model: e.target.value }))
-                                        }
-                                      }}
-                                      style={{
-                                        background: "#09090b", border: "1px solid rgba(63,63,70,0.6)",
-                                        borderRadius: "6px", padding: "0.45rem 0.6rem",
-                                        color: "#fff", fontSize: "0.8rem", fontFamily: "monospace",
-                                        cursor: "pointer",
-                                      }}
-                                    >
-                                      {PROVIDER_DEFAULTS[llmDraft.provider].models.map(m => (
-                                        <option key={m.id} value={m.id}>{m.label}</option>
-                                      ))}
-                                      <option value="__custom__" disabled style={{ color: "rgba(255,255,255,0.4)" }}>
-                                        ── custom (type below) ──
-                                      </option>
-                                    </select>
-                                    {/* Custom model input — shown when selection isn't in the list */}
-                                    <input
-                                      value={llmDraft.model}
-                                      onChange={e => setLLMDraft(d => ({ ...d, model: e.target.value }))}
-                                      placeholder="or type a custom model ID"
-                                      style={{
-                                        background: "#09090b", border: "1px solid rgba(63,63,70,0.4)",
-                                        borderRadius: "6px", padding: "0.4rem 0.6rem",
-                                        color: "rgba(255,255,255,0.7)", fontSize: "0.75rem", fontFamily: "monospace",
-                                      }}
-                                    />
-                                  </div>
-
-                                  {/* API key input */}
-                                  <div style={{ display: "flex", flexDirection: "column", gap: "0.35rem" }}>
-                                    <label style={{ fontSize: "0.75rem", color: "rgba(255,255,255,0.5)" }}>
-                                      API Key
-                                      {llmDraft.provider === "groq" && (
-                                        <span style={{ marginLeft: "0.4rem", color: "rgba(134,239,172,0.7)" }}>(free key at console.groq.com)</span>
-                                      )}
-                                    </label>
-                                    <input
-                                      type="password"
-                                      value={llmDraft.apiKeys[llmDraft.provider] ?? ""}
-                                      onChange={e => setLLMDraft(d => ({
-                                        ...d,
-                                        apiKeys: { ...d.apiKeys, [d.provider]: e.target.value }
-                                      }))}
-                                      placeholder={llmDraft.provider === "groq" ? "gsk_... (leave blank for env var)" : "Paste your API key"}
-                                      style={{
-                                        background: "#09090b", border: "1px solid rgba(63,63,70,0.6)",
-                                        borderRadius: "6px", padding: "0.45rem 0.6rem",
-                                        color: "#fff", fontSize: "0.8rem", fontFamily: "monospace",
-                                      }}
-                                    />
-                                    <span style={{ fontSize: "0.7rem", color: "rgba(255,255,255,0.3)" }}>
-                                      Stored in localStorage · never sent to FluidMCP servers
-                                    </span>
-                                  </div>
-
-                                  <div style={{ display: "flex", justifyContent: "flex-end", gap: "0.5rem" }}>
-                                    <button
-                                      onClick={() => setLLMSelectorOpen(false)}
-                                      style={{ fontSize: "0.75rem", padding: "0.35rem 0.75rem", borderRadius: "6px", background: "transparent", border: "1px solid rgba(63,63,70,0.5)", color: "rgba(255,255,255,0.5)", cursor: "pointer" }}
-                                    >
-                                      Cancel
-                                    </button>
-                                    <button
-                                      onClick={() => { saveLLMSettings(llmDraft); setLLMSelectorOpen(false); }}
-                                      style={{ fontSize: "0.75rem", padding: "0.35rem 0.75rem", borderRadius: "6px", background: "rgba(99,102,241,0.8)", border: "none", color: "#fff", cursor: "pointer", fontWeight: 600 }}
-                                    >
-                                      Save
-                                    </button>
-                                  </div>
-                                </div>
-                              </div>
-                            )}
-
-                            {/* ── Message input row ── */}
-                          <div style={{ display: "flex", gap: "0.5rem" }}>
-                            <input
-                              value={chatInput}
-                              disabled={chatLoading}
-                              onChange={(e) => setChatInput(e.target.value)}
-                              placeholder="Ask something..."
-                              onKeyDown={(e) => {
-                                if (e.key === "Enter" && !e.shiftKey && chatInput.trim()) {
-                                  e.preventDefault();
-                                  runChatTool();
-                                }
-                              }}
-                              style={{
-                                flex: 1, minWidth: 0, padding: "0.5rem 0.75rem",
-                                borderRadius: "0.5rem", border: "1px solid rgba(63,63,70,0.6)",
-                                background: "rgba(24,24,27,0.8)", color: "#fff", opacity: chatLoading ? 0.6 : 1,
-                                fontSize: "0.875rem",
-                              }}
-                            />
-                            <button
-                              onClick={()=>{ if (chatInput.trim()) runChatTool(); }}
-                              disabled={chatLoading || !chatInput.trim()}
-                              style={{
-                                padding: "0.5rem 1rem", borderRadius: "0.5rem",
-                                background: "rgba(99,102,241,0.9)", color: "#fff", fontWeight: "600", flexShrink: 0,
-                                border: "none",
-                                opacity: chatLoading || !chatInput.trim() ? 0.45 : 1,
-                                cursor: chatLoading || !chatInput.trim() ? "not-allowed" : "pointer",
-                                fontSize: "0.875rem",
-                              }}
-                            >
-                              {chatLoading ? <ThinkingDots /> : "Send"}
-                            </button>
-                          </div>
-                          </div>
-                        </>
-                      ) : (
-                        <div style={{
-                          padding: "1rem", border: "1px dashed rgba(239,68,68,0.6)",
-                          borderRadius: "0.5rem", textAlign: "center", color: "rgba(239,68,68,0.8)",
-                        }}>
-                          Server is not connected. Please reconnect to chat.
-                        </div>
-                      )}
-                    </div>
+                    <ChatPanel
+                      selectedServer={selectedServer}
+                      selectedServerId={selectedServerId!}
+                      chatHistory={chatHistory}
+                      executionHistory={executionHistory}
+                      chatInput={chatInput}
+                      chatLoading={chatLoading}
+                      systemPrompt={systemPrompt}
+                      systemPromptDraft={systemPromptDraft}
+                      systemPromptOpen={systemPromptOpen}
+                      chatRef={chatRef}
+                      chatBottomRef={chatBottomRef}
+                      setChatInput={setChatInput}
+                      setChatHistoryByServer={setChatHistoryByServer}
+                      setSystemPrompt={setSystemPrompt}
+                      setSystemPromptDraft={setSystemPromptDraft}
+                      setSystemPromptOpen={setSystemPromptOpen}
+                      onSendMessage={runChatTool}
+                      onSendMessageWithText={runChatToolWithMessage}
+                    />
                   )}
 
                   {/* ── RESOURCES MODE ─── */}
@@ -2533,10 +1252,11 @@ export default function MCPInspector() {
           customName={customName} setCustomName={setCustomName}
           envVars={envVars} setEnvVars={setEnvVars}
           recentUrls={recentUrls}
-          onClose={() => { setShowAddModal(false); setCustomName(""); setUrl(""); setCommand(""); setEnvVars([]); setTransport("http"); setAuthType("none"); setToken(""); setHeaderKey(""); setHeaderValue(""); }}
           onConnect={handleConnect}
+          onClose={() => { setShowAddModal(false); setCustomName(""); setUrl(""); setCommand(""); setEnvVars([]); setTransport("http"); setAuthType("none"); setToken(""); setHeaderKey(""); setHeaderValue(""); }}
         />
       )}
+
     </div>
   );
 }
